CORE: Portability fixes, plus the addition of x64 platform for Windows builds (not functional yet, needs work in src\common\zlib.cpp before 64-bit builds are possible).
64-bit windows dynamic libraries added.

64-bit VS builds require either Windows SDK 7.1 with the 64-bit compilers installed or VS 2012. It is configured for Windows SDK 7.1 by default.

Felt the need to commit this as-is because I've spent far too much time trying to work out the final kinks. I need to catch up with the updates to the code-base now before continuing.

+++ src/common/taskmgr.cpp
@@ -55,7 +55,7 @@ void CTaskMgr::RemoveTask(std::string TaskName)
 
 uint32 CTaskMgr::DoTimer(uint32 tick)
 {
-	int32 diff = 1000; 
+	size_t diff = 1000; 
 
 	while( !m_TaskList.empty() )
 	{
